[ALPS05092356] FPSGO: add FPSGO driver
[Detail] add FPSGO driver

MTK-Commit-Id: 8217cc5704f1036ed2c734c527882e67f59c2f1c

CR-Id: ALPS05092356
Feature: FPSGO

+++ drivers/misc/mediatek/performance/fpsgo_v3/composer/include/fps_composer.h
@@ -19,7 +19,7 @@ struct connect_api_info {
 	int pid;
 	int tgid;
 	unsigned long long buffer_id;
-	int buffer_key;
+	unsigned long long buffer_key;
 	int api;
 };
